Update config page and a single ricebox
