--- mgmt-js.rice-box.go
@@ -1,41 +1,0 @@
-package main
-
-import (
-	"github.com/GeertJohan/go.rice/embedded"
-	"time"
-)
-
-func init() {
-
-	// define files
-	filea := &embedded.EmbeddedFile{
-		Filename:    `mgmt.js`,
-		FileModTime: time.Unix(1438577974, 0),
-		Content:     string([]byte{0x76, 0x61, 0x72, 0x20, 0x4d, 0x67, 0x6d, 0x74, 0x20, 0x3d, 0x20, 0x4d, 0x67, 0x6d, 0x74, 0x20, 0x7c, 0x7c, 0x20, 0x7b, 0x7d, 0x3b, 0xa, 0xa, 0x76, 0x61, 0x72, 0x20, 0x73, 0x68, 0x6f, 0x77, 0x48, 0x69, 0x64, 0x65, 0x20, 0x3d, 0x20, 0x66, 0x75, 0x6e, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x20, 0x28, 0x29, 0x20, 0x7b, 0xa, 0x20, 0x20, 0x20, 0x20, 0x24, 0x28, 0x22, 0x2e, 0x68, 0x69, 0x64, 0x64, 0x65, 0x6e, 0x22, 0x29, 0x2e, 0x68, 0x69, 0x64, 0x65, 0x28, 0x29, 0x3b, 0xa, 0x7d, 0x3b, 0xa, 0xa, 0x76, 0x61, 0x72, 0x20, 0x74, 0x6f, 0x67, 0x67, 0x6c, 0x65, 0x53, 0x68, 0x6f, 0x77, 0x20, 0x3d, 0x20, 0x66, 0x75, 0x6e, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x28, 0x65, 0x6c, 0x65, 0x6d, 0x29, 0x7b, 0xa, 0x20, 0x20, 0x69, 0x66, 0x20, 0x28, 0x24, 0x28, 0x65, 0x6c, 0x65, 0x6d, 0x29, 0x2e, 0x68, 0x61, 0x73, 0x43, 0x6c, 0x61, 0x73, 0x73, 0x28, 0x27, 0x68, 0x69, 0x64, 0x64, 0x65, 0x6e, 0x27, 0x29, 0x29, 0x20, 0x7b, 0xa, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x24, 0x28, 0x65, 0x6c, 0x65, 0x6d, 0x29, 0x2e, 0x72, 0x65, 0x6d, 0x6f, 0x76, 0x65, 0x43, 0x6c, 0x61, 0x73, 0x73, 0x28, 0x27, 0x68, 0x69, 0x64, 0x64, 0x65, 0x6e, 0x27, 0x29, 0x3b, 0xa, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x24, 0x28, 0x65, 0x6c, 0x65, 0x6d, 0x29, 0x2e, 0x73, 0x68, 0x6f, 0x77, 0x28, 0x29, 0x3b, 0xa, 0x20, 0x20, 0x7d, 0x20, 0x65, 0x6c, 0x73, 0x65, 0x20, 0x7b, 0xa, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x24, 0x28, 0x65, 0x6c, 0x65, 0x6d, 0x29, 0x2e, 0x61, 0x64, 0x64, 0x43, 0x6c, 0x61, 0x73, 0x73, 0x28, 0x27, 0x68, 0x69, 0x64, 0x64, 0x65, 0x6e, 0x27, 0x29, 0x3b, 0xa, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x24, 0x28, 0x65, 0x6c, 0x65, 0x6d, 0x29, 0x2e, 0x68, 0x69, 0x64, 0x65, 0x28, 0x29, 0x3b, 0xa, 0x20, 0x20, 0x7d, 0xa, 0x7d, 0x3b, 0xa, 0x4d, 0x67, 0x6d, 0x74, 0x2e, 0x69, 0x6e, 0x69, 0x74, 0x69, 0x61, 0x6c, 0x69, 0x7a, 0x65, 0x20, 0x3d, 0x20, 0x66, 0x75, 0x6e, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x20, 0x28, 0x29, 0x20, 0x7b, 0xa, 0x20, 0x20, 0x20, 0x20, 0x63, 0x6f, 0x6e, 0x73, 0x6f, 0x6c, 0x65, 0x2e, 0x6c, 0x6f, 0x67, 0x28, 0x22, 0x49, 0x6e, 0x69, 0x74, 0x69, 0x61, 0x6c, 0x69, 0x7a, 0x65, 0x64, 0x22, 0x29, 0x3b, 0xa, 0x20, 0x20, 0x20, 0x20, 0x73, 0x68, 0x6f, 0x77, 0x48, 0x69, 0x64, 0x65, 0x28, 0x29, 0x3b, 0xa, 0x20, 0x20, 0x20, 0x20, 0x24, 0x28, 0x22, 0x2e, 0x73, 0x68, 0x6f, 0x77, 0x2d, 0x6f, 0x69, 0x64, 0x73, 0x22, 0x29, 0x2e, 0x6f, 0x6e, 0x28, 0x22, 0x63, 0x6c, 0x69, 0x63, 0x6b, 0x22, 0x2c, 0x20, 0x66, 0x75, 0x6e, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x20, 0x28, 0x65, 0x6c, 0x65, 0x6d, 0x29, 0x20, 0x7b, 0xa, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x63, 0x6f, 0x6e, 0x73, 0x6f, 0x6c, 0x65, 0x2e, 0x6c, 0x6f, 0x67, 0x28, 0x22, 0x43, 0x6c, 0x69, 0x63, 0x6b, 0x65, 0x64, 0x22, 0x29, 0x3b, 0xa, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x63, 0x6f, 0x6e, 0x73, 0x6f, 0x6c, 0x65, 0x2e, 0x6c, 0x6f, 0x67, 0x28, 0x65, 0x6c, 0x65, 0x6d, 0x29, 0x3b, 0xa, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x76, 0x61, 0x72, 0x20, 0x70, 0x20, 0x3d, 0x20, 0x24, 0x28, 0x65, 0x6c, 0x65, 0x6d, 0x2e, 0x74, 0x61, 0x72, 0x67, 0x65, 0x74, 0x29, 0x3b, 0xa, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x74, 0x6f, 0x67, 0x67, 0x6c, 0x65, 0x53, 0x68, 0x6f, 0x77, 0x28, 0x24, 0x28, 0x70, 0x29, 0x2e, 0x63, 0x6c, 0x6f, 0x73, 0x65, 0x73, 0x74, 0x28, 0x22, 0x74, 0x72, 0x22, 0x29, 0x2e, 0x66, 0x69, 0x6e, 0x64, 0x28, 0x22, 0x2e, 0x6f, 0x69, 0x64, 0x73, 0x22, 0x29, 0x29, 0x3b, 0xa, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x20, 0x74, 0x6f, 0x67, 0x67, 0x6c, 0x65, 0x53, 0x68, 0x6f, 0x77, 0x28, 0x24, 0x28, 0x70, 0x29, 0x2e, 0x63, 0x6c, 0x6f, 0x73, 0x65, 0x73, 0x74, 0x28, 0x22, 0x74, 0x72, 0x22, 0x29, 0x2e, 0x66, 0x69, 0x6e, 0x64, 0x28, 0x22, 0x2e, 0x6f, 0x69, 0x64, 0x2d, 0x68, 0x69, 0x64, 0x64, 0x65, 0x6e, 0x22, 0x29, 0x29, 0x3b, 0xa, 0x20, 0x20, 0x20, 0x20, 0x7d, 0x29, 0x3b, 0xa, 0x7d, 0x3b, 0xa, 0xa, 0x77, 0x69, 0x6e, 0x64, 0x6f, 0x77, 0x2e, 0x6f, 0x6e, 0x6c, 0x6f, 0x61, 0x64, 0x20, 0x3d, 0x20, 0x66, 0x75, 0x6e, 0x63, 0x74, 0x69, 0x6f, 0x6e, 0x20, 0x28, 0x29, 0x20, 0x7b, 0xa, 0x20, 0x20, 0x20, 0x20, 0x4d, 0x67, 0x6d, 0x74, 0x2e, 0x69, 0x6e, 0x69, 0x74, 0x69, 0x61, 0x6c, 0x69, 0x7a, 0x65, 0x28, 0x29, 0x3b, 0xa, 0x7d, 0x3b, 0xa}), //++ TODO: optimize? (double allocation) or does compiler already optimize this?
-	}
-
-	// define dirs
-	dir9 := &embedded.EmbeddedDir{
-		Filename:   ``,
-		DirModTime: time.Unix(1438577974, 0),
-		ChildFiles: []*embedded.EmbeddedFile{
-			filea, // mgmt.js
-
-		},
-	}
-
-	// link ChildDirs
-	dir9.ChildDirs = []*embedded.EmbeddedDir{}
-
-	// register embeddedBox
-	embedded.RegisterEmbeddedBox(`mgmt/js`, &embedded.EmbeddedBox{
-		Name: `mgmt/js`,
-		Time: time.Unix(1438578538, 0),
-		Dirs: map[string]*embedded.EmbeddedDir{
-			"": dir9,
-		},
-		Files: map[string]*embedded.EmbeddedFile{
-			"mgmt.js": filea,
-		},
-	})
-}
