+++ mgmt/templates/config.tmpl
@@ -1,27 +1,78 @@
 <div class="container">
-  <p><strong>URL:</strong> {{.Config.Scheme}}://{{.Config.Host}}</p>
-  <p><strong>Listen Address:</strong> {{.Config.Listen}}</p>
-  <p><strong>Meta Storage Type:</strong> {{.Config.BackingStore}}</p>
-  <p><strong>Database:</strong> {{.Config.MetaDB}}</p>
-  <p><strong>Content Store Type:</strong> {{.Config.ContentStore}}</p>
-  <p><strong>Content Path:</strong> {{.Config.ContentPath}}</p>
-  <p><strong>Public:</strong> {{.Config.Public}}</p>
+  <p>To configure a repository to use this LFS server, add the following to the repository's <code>.gitconfig</code> file:</p>
+  <pre>
+  <code>[lfs]
+      url = "{{.Config.Scheme}}://{{.Config.Host}}/&lt;user&gt;/repo"
+  </code>
+  </pre>
+
+  {{if eq .Config.Scheme "https"}}
+  <p>Your server is configured to use https. If you're using self signed certificates, or are getting SSL errors, you can add the following to your <code>.gitconfig</code> file:</p>
+  <pre>
+  <code>[http]
+      sslverify = false
+  </code>
+  </pre>
+{{end}}
 </div>
 <br>
 <div class="container">
-  <p>To configure a repository to use this LFS server, add the following to the repository's <code>.gitconfig</code> file:</p>
-  <pre>
-<code>[lfs]
-    url = "{{.Config.Scheme}}://{{.Config.Host}}/user/repo"
-</code>
-</pre>
-
-{{if eq .Config.Scheme "https"}}
-<p>Your server is configured to use https. If you're using self signed certificates, or are getting SSL errors, you can add the following to your <code>.gitconfig</code> file:</p>
-<pre>
-<code>[http]
-    sslverify = false
-</code>
-</pre>
+  <table>
+      <tr>
+        <td><h2>General Config</h2></td>
+      </tr>
+      <tr>
+        <td><strong>URL:</strong></td>
+        <td>{{.Config.Scheme}}://{{.Config.Host}}</td>
+      </tr>
+      <tr>
+          <td><strong>Listen Address:</strong></td>
+          <td>{{.Config.Listen}}</td>
+      </tr>
+      <tr>
+          <td><strong>Meta Storage Type:</strong></td>
+          <td>{{.Config.BackingStore}}</td>
+      </tr>
+      <tr>
+          <td><strong>Database:</strong></td>
+          <td>{{.Config.MetaDB}}</td>
+      </tr>
+      <tr>
+          <td><strong>Content Store Type:</strong></td>
+          <td>{{.Config.ContentStore}}</td>
+      </tr>
+      <tr>
+          <td><strong>Content Path:</strong></td>
+          <td>{{.Config.ContentPath}}</td>
+      </tr>
+      <tr>
+          <td><strong>Public:</strong></td>
+          <td>{{.Config.Public}}</td>
+      </tr>
+      <tr>
+          <td><strong>GoProcs:</strong></td>
+          <td>{{.Config.NumProcs}}</td>
+      </tr>
+  </table>
+  {{if .Config.Ldap.Enabled}}
+    <table>
+      <tr>
+          <td>LDAP Config</td>
+      </tr>
+    </table>
+  {{end}}
+  {{if .Config.Cassandra.Enabled}}
+    <table>
+      <tr>
+          <td>Cassandra Config</td>
+      </tr>
+    </table>
+  {{end}}
+  {{if .Config.Aws.Enabled}}
+    <table>
+      <tr>
+          <td>AWS Config</td>
+      </tr>
+    </table>
+  {{end}}
 </div>
-{{end}}
